sound/nes_apu.cpp: Added earliest hardware variant of 2A03 APU. (#10299)
Fixes several audio bugs with the noise channel in VS. System games, including:
* High-pitch sound in vsgshoe percussion track.
* Jet sounds in bnglngby.
* Number of tanks killed count screen in btlecity.
* nvs_platoon's bullet and enemies dying sounds are subtly changed.

It's estimated that these letterless CPU versions are in the first two
million or so Famicoms (about 10% of Famicoms sold).  There are games
developed on the RP2A03 that have sound glitches on later, more common
Famicoms.  For instance, the very last note in the game over melody in
Balloon Fight rings out with a loud buzz on later machines.  The new
famicomo driver plays it as the developers must have intended.

New working clones
------------------
Nintendo Famicom (earlier, with RP2A03) [kmg]